Move the definition of a branch to its own file
